System_V2/cam_ui/requirements.txt
grpcio==1.36.1
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.2
protobuf==3.15.6
pypylon==1.7.2
six==1.15.0
